+++ relay-server/src/services/processor.rs
@@ -18,7 +18,7 @@ use fnv::FnvHasher;
 use relay_base_schema::project::{ProjectId, ProjectKey};
 use relay_cogs::{AppFeature, Cogs, FeatureWeights, ResourceId, Token};
 use relay_common::time::UnixTimestamp;
-use relay_config::{Config, HttpEncoding, RelayMode};
+use relay_config::{Config, HttpEncoding, Normalize, RelayMode};
 use relay_dynamic_config::{ErrorBoundary, Feature};
 use relay_event_normalization::{
     normalize_event, validate_event_timestamps, validate_transaction, ClockDriftProcessor,
@@ -1210,6 +1210,25 @@ impl EnvelopeProcessorService {
         &self,
         state: &mut ProcessEnvelopeState<G>,
     ) -> Result<(), ProcessingError> {
+        let full_normalization = match self.inner.config.normalization() {
+            Normalize::Disabled => {
+                // We assume envelopes coming from an internal relay have
+                // already been normalized. During incidents, like a PoP region
+                // not being available, envelopes can go to other PoP regions or
+                // directly to processing relays. Events should be fully
+                // normalized, independently of the ingestion path.
+                if self.inner.config.processing_enabled()
+                    && !state.envelope().meta().is_from_internal_relay()
+                {
+                    true
+                } else {
+                    return Ok(());
+                }
+            }
+            Normalize::Full => true,
+            Normalize::Default => self.inner.config.processing_enabled(),
+        };
+
         if let Some(sampling_state) = state.sampling_project_state.clone() {
             state
                 .envelope_mut()
@@ -1240,13 +1259,11 @@ impl EnvelopeProcessorService {
                 is_validated: false,
             };
 
-            let is_last_relay_normalize = self.inner.config.processing_enabled();
-
             let key_id = state
                 .project_state
                 .get_public_key_config()
                 .and_then(|key| Some(key.numeric_id?.to_string()));
-            if is_last_relay_normalize && key_id.is_none() {
+            if full_normalization && key_id.is_none() {
                 relay_log::error!(
                     "project state for key {} is missing key id",
                     state.managed_envelope.envelope().meta().public_key()
@@ -1292,8 +1309,8 @@ impl EnvelopeProcessorService {
                     .aggregator_config_for(MetricNamespace::Spans)
                     .max_tag_value_length,
                 is_renormalize: false,
-                remove_other: is_last_relay_normalize,
-                emit_event_errors: is_last_relay_normalize,
+                remove_other: full_normalization,
+                emit_event_errors: full_normalization,
                 span_description_rules: state.project_state.config.span_description_rules.as_ref(),
                 geoip_lookup: self.inner.geoip_lookup.as_ref(),
                 enable_trimming: true,
@@ -1315,7 +1332,7 @@ impl EnvelopeProcessorService {
                 validate_transaction(event, &tx_validation_config)
                     .map_err(|_| ProcessingError::InvalidTransaction)?;
                 normalize_event(event, &normalization_config);
-                if is_last_relay_normalize && event::has_unprintable_fields(event) {
+                if full_normalization && event::has_unprintable_fields(event) {
                     metric!(counter(RelayCounters::EventCorrupted) += 1);
                 }
                 Result::<(), ProcessingError>::Ok(())