+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 - Tag span duration metric by group for all ops supporting description scrubbing. ([#3370](https://github.com/getsentry/relay/pull/3370))
 - Copy transaction tags to segment. ([#3386](https://github.com/getsentry/relay/pull/3386))
 - Route spans according to trace_id. ([#3387](https://github.com/getsentry/relay/pull/3387))
+- Optionally skip normalization. ([#3377](https://github.com/getsentry/relay/pull/3377))
 
 ## 24.3.0